refactor: extract SocialLinks component
