+++ components/AboutMe.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import userData from "../constants/data";
+import SocialLinks from "./SocialLinks";
 
 const AboutMe: React.FC = () => {
     return (
@@ -58,44 +59,7 @@ const AboutMe: React.FC = () => {
                         <h2 className="text-xl font-semibold text-gray-700 mt-8 dark:text-gray-200">
                             Social Links
                         </h2>
-                        <div className="mt-4 ml-4">
-                            <div className="flex flex-row justify-start items-center">
-                                <a
-                                    href={userData.socialLinks.linkedin}
-                                    className="flex flex-row items-center space-x-4 group"
-                                >
-                                    <div className="my-4">&rarr;</div>
-                                    <div className="text-lg text-gray-500 font-mono relative overflow-hidden dark:text-gray-300">
-                                        <div className="absolute h-0.5 w-full bg-gray-400 bottom-0 transform -translate-x-24 group-hover:translate-x-0 transition duration-300"></div>
-                                        LinkedIn
-                                    </div>
-                                </a>
-                            </div>
-                            <div className="flex flex-row justify-start items-center">
-                                <a
-                                    href={userData.socialLinks.instagram}
-                                    className="flex flex-row items-center space-x-4 group"
-                                >
-                                    <div className="my-4">&rarr;</div>
-                                    <div className="text-lg text-gray-500 font-mono relative overflow-hidden dark:text-gray-300">
-                                        <div className="absolute h-0.5 w-full bg-gray-400 bottom-0 transform -translate-x-28 group-hover:translate-x-0 transition duration-300"></div>
-                                        Instagram
-                                    </div>
-                                </a>
-                            </div>
-                            <div className="flex flex-row justify-start items-center">
-                                <a
-                                    href={userData.socialLinks.twitter}
-                                    className="flex flex-row items-center space-x-4 group"
-                                >
-                                    <div className="my-4">&rarr;</div>
-                                    <div className="text-lg text-gray-500 font-mono relative overflow-hidden dark:text-gray-300">
-                                        <div className="absolute h-0.5 w-full bg-gray-400 bottom-0 transform -translate-x-24 group-hover:translate-x-0 transition duration-300"></div>
-                                        Twitter
-                                    </div>
-                                </a>
-                            </div>
-                        </div>
+                            <SocialLinks className="mt-4 ml-4" />
                     </div>
                     {/* Text area */}
                     <div className="col-span-1 md:col-span-2">
